So after many hours of image editing and tweaking I bring to you "DUN DUN DUN" help screens!
I modified some of the code in MarketScreen because of an unexpected behavior.
Apparently removing a frame from the scene does not remove its button press
procedures...yes I was able to click buttons without even having the screen displayed.

Lastly, fixed a small issue with the way the market and inventory were used
so that we didn't recreate the objects over and over and over.  It also helps
a lot to keep it this way because we don't need any static variables to track
when a player has already seen a help screen.  And no, I refused to put variables
for that task into ARRG.cs :-P

I still need to do that ARRG title screen background.  Maybe if I wake up early enough...